requirements.txt
PyPDF2
pdf2image
pdfplumber
Pillow